Add Experimental annotation for KeyInput
The KeyInput API is experimental and is likely to change in the future. This CL adds an experimental annotation that can be added to all key input related API. Users will have to @OptIn to use key input. We will remove the annotation when the API is stable.

Bug: 160907717
Relnote: Added an experimental api annotation for key input related API
Test: Built and launched the demo app
Change-Id: I53c0a52af14dd832170f774cbb638c253abd639d