Release Surface from ImageReader
Release the Surface from ImageReader that is used for ImageAnalysis and
ImageCapture. They were not released, which was causing a StrictMode
violation on API levels 21/22.

The Surface is release as part of the SafeCloseImageReader which makes
sure that the Surface is release when the ImageReaderProxy is closed.
This is safe since the Surface should no longer be used once the
ImageProxy has been closed.

Bug: b/157162060
Test: ./gradlew camera:camera-core:{test,connectedCheck}

Change-Id: I968d2ba5280c9d2b468842ef59608f4ad9ecca22