Fixes benchmark APKs not having unique names
Previously parent projects were not included in the output file name for benchmarks, meaning that:

foo:benchmark
bar:benchmark

Would produce the same output APK, causing a clash and one to be overwritten. This CL prepends the full path, just as we do
for non-benchmark APKs. The media setup is complex and specific, so leaving that alone in this CL.

Test: ./gradlew buildTestApks, verify that there are unique apks produced
Bug: b/161157211
Change-Id: Ibce5db10ec58bd10832357392698116a0ebf39d5