Fix popup placement with cut-outs.
Popup was supplying screen relative position to window manager which
needs a window relative one. This fix shifts the position and also
provides a proper window coordinates that can be used to position the
popup in custom position providers.

Also fixed a leak that happens when isFocusable gets changes.

Bug: 157657482
Test: Added
Relnote: Fixed popup position on cut-out displays.
Change-Id: Idd7ddf5f88728a96fc8ff725a39979fe962e8889