Update AndroidView interop APIs
This CL:
- adds a new API for composing an Android View, which takes a lambda
creating the View. This API also takes an update block, which
observes state and will update the View when this changes
- deprecates the previously existing AndroidView overloads. Both can be
replaced with the new API. In addition, the XML inflation AndroidView
will have (in a follow up CL) a ViewBinding based correspondent which
will be the recommended replacement.

Relnote: Added a new AndroidView API and deprecated the existing ones.
Fixes: 161704511
Fixes: 161705010
Change-Id: I5bcfd9c9945fecb3e55fe95d6528d8391ac7b961