Move PresenterCallback into PagingDataDiffer
First step towards the final public API for PagingDataDiffer,
essentially it needs to accept a diffing implementation for REFRESH, and
output ListUpdateCallback calls for incremental updates.

This change allows initial refresh to be synchronous with presenter
updates, as PagingDataDiffer now has the ability to hold off on
dispatching LoadState updates until the presenter has been fully updated
and set.

Fixes: 160762353
Test: ./gradlew paging:paging-runtime:cC
Change-Id: I5d32a47984b7eae8330fb13665864ab942929070