Introduce AnimatedVisibility
Relnote: "AnimatedVisibilty composable animates the appearance
and disappearance of the child content.

EnterTransition and ExitTransition are introduced to work with
AnimatedVisibilty composable to provide 3 different typs of
appearance and disappearance animation: fade, slide, and
expand/shrink the content. The different types of animations
can be combined to achieve more bespoke look and feel.
"

Bug: 147610583
Bug: 147610432
Test: unit tests added
Change-Id: Idda1162e0d7f777f64dfd91b3192c2548d1c2c29