Find and subscribe Compose Preview animations on layout
Use the slot table to find transition animations on the user code and
subscribe the corresponding AnimationClockObservers to the Compose Preview's
PreviewAnimationClock. This way, we don't need to wait for animations to be
subscribed to keep track of them on Android Studio's Animation Inspector.

For now, only transition animations are found, but we're planning to extend
this behavior for other animations (e.g. single animated value) in future
versions of the animation inspector.

Bug: 160747435
Test: ComposeViewAdapterTest#transitionAnimationsAreSubscribedToTheClock
Change-Id: I0f9e18394050a7b709f379ea3a1c23a75f96288f