Invalidate SurfaceRequest if new SurfaceTexture is available
When the camera preview is using a TextureView, when a new SurfaceTexture becomes available, if the TextureView is already streaming a camera preview, the surfaceRequest has to be invalidated so that a new Surface is provided to the camera for preview.

This issue was discovered from the test PreviewViewFragmentTest#checkPreviewUpdatedAfterToggleCameraAndStopResume(), after the lifecycle's onstop then onStart event, a new SurfaceTexture was created (on certain devices), but the camera was still holding on to a Surface from the previous SurfaceTexture, which was resulting in a blank screen.

Bug: 155195589
Test: PreviewViewFragmentTest.java
Change-Id: Ia70548a9aad0368642b778621979e7edadbb4b19