Ensure fragment views are re-added to the container
When using the new move to state (USE_STATE_MANAGER=true), if a fragment
has started animating out its view is removed from the container. If
the fragment is re-added to the fragment manager before the animation
has a chance to finish, the exit animation is cancelled and the
fragment reuses the same view. This view should be added back to the
container or there can be animation issues.

Test: FragmentContainerViewTest pass in new world (8 tests failing)
Bug: 147749580
Change-Id: Ie267edc514c53260fd9219f8fa99171f831149d2