Wait for onCommits in ComposeIdlingResources
When we want to know if Compose is idle, we post a message to the main
handler. When it is executed, isIdle() reports either true or false. If
it reports false, repost the message and check again.

If isIdle() reports true, there is no pending composition, and there can
either be onCommits scheduled on the Choreographer or not. Either way,
schedule a message on the Choreographer.

When it is executed, and isIdle() now reports false, restart from the
beginning. If isIdle() reports true, continue.

If there were no onCommits scheduled, nothing changed in between and we
know Compose is idle. Stop and inform the IdlingResource. If there were
onCommits scheduled, they have been executed by now. If they triggered a
recomposition, it is now pending and isIdle() returned false. If they
didn't, we know Compose is now idle. Stop and inform the IdlingResource.

Fixes: 160399857
Bug: 160326855
Test: ./gradlew ui:ui-test:cC \
        compose:material:material:cC \
        compose:foundation:foundation:cC \
        compose:ui:ui:cC
Change-Id: I385afb23af01ff8d839f9ed187b2e45bcbbfcaf7