Remove options extra from fillInIntent
We need to start the ActivityOptions in the fillInIntent to ensure they
get passed to the activity to make it in the actual call for
startIntentSender for result. After the options are delivered to the
FragmentIntentSenderContract, we should remove that extra from the
fillInIntent and remove the fillInIntent itself if it was created only
to pass the extra

This changed also cleans some unused code.

Test: verified through debugging FragmentReceiveResultTest
Bug: 159805984
Change-Id: I7c5c060d32e435d7264fbf5ea2e5891056291b87