Avoid calling execute in layout / draw passes
The previous behavior of onStartEnterTransition()
would schedule a commit for the next frame, which
avoiding executing transitions in the middle of
a layout or draw pass (such as when the developer
calls startPostponedEnterTransition() from a
doOnPreDraw lambda).

We need to do the same post when calling
executePendingOperations() so as to avoid visual
artifacts (the fragments appearing in their final state
for a single frame before the transition begins).

By tracking whether callStartTransitionListener()
was called directly or whether it was posted, we
can ensure that we are calling markPostponedState()
as soon as possible and calling executePendingOperations()
as early as is safe.

Test: tested in sample app, existing tests pass
BUG: 162829603
Change-Id: I625d829ae89d6e613ef8cb85c3fc25ac5ce5d42e