Add overflow and resistance to Modifier.swipeable
This CL does the following things:
* Add support for observing the overflow or the amount by which the
composable has been swiped outside of its bounds using SwipeableState.
* Add resistance to Modifier.swipeable when swiping past the bounds.
Resistance is calculated by scaling the overflow using the size of the
composable as a basis, applying the sine function, and multiplying by a
resistance factor. There are different resistance factors at each edge,
and these are customisable as parameters. Two default resistance factors
are provided as constants: a standard one (10f) and a stiff one (20f).
* Remove minValue and maxValue from Modifier.swipeable, since resistance
eliminates the need to have bounds that are not the min and max anchors.

Test: Ran SwipeableTest and SwipeToDismissTest
Test: Added new tests for overflow and resistance in SwipeableTest
Relnote: "Added params in Modifier.swipeable for changing the amount of
resistance when swiping past the bounds. Removed [min/max]Value params."
Change-Id: I93d98d2674798d39844f88dcef356cbbdb34458a