Implement per-property merge policy.
This changes the current merging behavior into the default
implementation of an open function called "merge()", and adds different
implementations for the following:
- Text and Label: concatenates strings with comma in between (same as
  TalkBack's native merging behavior).  This behavior will probably
  evolve further in future patches to merge text layout for accessibility.
- TestTag and Hidden: does not merge at all.
- Popup and Dialog: throws an exception.

Note that the latter two behaviors require carefully distinguishing
between collapsing and merging so that, for example, collapsed TestTags
are not dropped as well.

Also switch ListItem to have mergeAllDescendants (should've had that
before anyway), which has more multiple-Text stress tests than
Button/Toggleable. As an example of how TalkBack behaves with a two-line
ListItem, see https://screenshot.googleplex.com/06KZaX6VfsG

Fixes: 161979921
Test: depthFirstLabelConcat, clickableTest_clickOnChildText
Relnote: "Added mergePolicy lambda to SemanticsPropertyKey.  This can be
used to define a custom policy for mergeAllDescendants semantics
merging.  The default policy is to use the parent value if already
present, otherwise the child value."

Change-Id: Iaf6c4cc327017ee492f4d8334c8df5167d33df58