Inject input events batched from the main thread
Instead of sleeping on the test thread to inject events more or less at
the same time as their eventTime says, generate all events up front and
dispatch them sequentially in a coroutine on the main dispatcher.

This rules out the possibility that thread contention can delay an event
so long that it is injected much later than its eventTime, which could
trigger other time based processes that wouldn't have triggered had the
event been injected on time.

Fixes: 163112000
Fixes: 158689874
Bug: 163112817
Test: ./gradlew ui:ui-test:cC compose:foundation:foundation:cC
Change-Id: Iddd840d29542bade2f23503e5aa3757c621cd4e4