TapGestureFilter that does not consume.
See bug for details as to why this needs
to exist.

TLDR: Multiple top level systems want to
know if the user tapped on an empty gesture
space in the app so that they each can
respond. The only way for each to respond
is to have a TapGestureFilter that does not
block TapGestureFilters. This is that.

Fixes: 163421610
Test: ./gradlew compose:ui:ui:test
Test: ./gradlew compose:ui:ui:connectedCheck

Change-Id: I1e99fe061eff2fa2f04e0e17ebd6b0ad5f3e3e11