extraCommand and extraCallbackWithResult back to Nullable
This CL partially revert aosp/1378188.

Revert the "Nullable->NonNull" part of the change because changing
these APIs to NonNull may break backward compatibility, and should
be defer to next major release.

But we do want to keep the "adding KEY_SUCCESS" part to provide a
way for browser and client indicate the command is success.
Since the "extraCommand" APIs are entirely implementation-defined, the
client and browser should decide whether use "KEY_SUCCESS" or not.

Test: no need
Relnote: "Revert change 'Makes CustomTabsService#extraCommand,
    CustomTabsCallback#extraCallbackWithResult and
      TrustedWebActivityService#onExtraCommand returns @NonNull'

    Adds CustomTabsService#KEY_SUCCESS and
      TrustedWebActivityService#KEY_SUCCESS that can be use for
      indicating extraCommand is success."

Change-Id: I6f7b55d173b4df07e742ba44e3f4b41306fd3640