Add support for setting the ShareState.
This allows developers to choose whether to show a share option or not (or leave it up to the browser).
This also removes the setDefaultShareMenuItemEnabled() method in favor of the new setShareState() method.

Test: Added tests to CustomTabsIntentTest.
Relnote: "Adds CustomTabsIntent#setShareState, which allows developers to specify whether to show a share option or not (or leave it up to the browser)."

Change-Id: I153fed30ea5814474c76dada550f49abede472f3