MediaRouter: Adjust volume correctly
GroupRouteController#onUpdateVolume used RoutingController#getVolume
as the previous volume, which might not be updated yet.
As a result, holding a physical volume button to increase/decrease
volume didn't work correctly i.e., the volume bar changed after releasing the button.

This CL changes GroupRouteController to remember lastly set volume
to adjust volume correctly.
The logic is the same as MediaSessionRecord#adjustVolume.

Bug: 160205155
Test: Run support v7 demos and cast to a Chromecast device, then
hold a physical volume button and release to increase/decrease the volume.

Change-Id: I5c79ed2731be0ee7cab3cc2efc79547bcc9e37a0