Deprecate LiveData.observe Kotlin extension
With the move to Kotlin 1.4, Kotlin itself
is able to convert the default observe() to the
lambda syntax previously only available when
using the Kotlin extension. To avoid users seeing
"Candidate resolution will be changed soon" errors,
we should deprecate our Kotlin extension as it
is no longer needed.

Test: ./gradlew bOS
Relnote: "The `LiveData.observe()` Kotlin extension
necessary to use lambda syntax is now deprecated as it
is not necessary when using Kotlin 1.4."

Change-Id: I40d3f40cbaebda0df2292d60704d9c19c9d835b9