Few fixes in jetifier.
1) Fix ignored tests after upgrade to kotlin 1.4
The tests were passing empty config property and the gson parser made
the property null which Kotlin did not like as the property was not
nullable. Not sure how this could work before so made it nullable as it
should be.
2) Removed warning for packages mismatch when dejetifying support lib.
This is fix for 166466059.
3) Removed throwing exception on missing package mapping when
dejetifying MetaInf files. Because we are now stable so this should be
ok. Also it was not very intuitive and folks were hitting this error for
new libraries all the time.

Fix: 165300826
Fix: 166466059
Test: Fixed
Change-Id: Icd948c07901f0320dd9f5bc26f56d2da0390ff10