Renamed a few things plus removed DrawMode.COMPLICATION_SELECT
The API council felt Complication.computeScreenBounds didn't fit on
the public API so I made it internal which meant we couldn't use it
in the watch face sample. This prompted me to move complication
select rendering into the library (and from there ultimatlty it
will go into SysUi) which is something I've considered doing for a
while to simplify the DrawModes, making it more likely watch face
developers will do the right thing.

Test: Example watch face worksas expected
Change-Id: Idee6a6bbd34d49b32075811ce8a3bf72762dd244