Fix CryptoObject type in BiometricManager reflective call
Ensures that an android.hardware.biometrics.BiometricPrompt.CryptoObject
is passed to the BiometricManager#canAuthenticate(CryptoObject) non-API
method, instead of androidx.biometric.BiometricPrompt.CryptoObject. Also
adds IllegalArgumentException to the list of handled exceptions for this
reflective invocation, to avoid crashing on a device with a different
method signature (since it's not part of the public Android API).

Test: Biometric integration test app on Samsung device with API 29.
Fixes: 165824669
Change-Id: I550d06679610ed423e8a4362a738f6ae4ff49279