Preview default size defined in camera2 PreviewConfigProvider
Moved the check to get default preview size in PreviewConfigProvider
since it is a camera2 implementation detail.

Remove CameraX.getSurfaceManager() singleton call
Remove need for ShadowCameraX since UseCases no longer using the CameraX
singleton methods.

Relnote: "Creation of UseCases do not require initialization of CameraX
to complete. All implementation specific configurations are set on
UseCase once it is attached to a Camera instance which for the public
API is ProcessCameraProvider.bindToLifecycle()."

Bug: b/162012861
Test: ./gradlew camera:camera-core:test
Change-Id: Ia5411289c68c202dae2316b8a9987d6d28a3d749