heifwriter: make sure shutdown doesn't hang or crash
Previous change to wait for countdown latch during shutdown turns
out to be problematic, it coud hang if writer is stopped without
start. This also adds delay to apps and is not necessary, as apps
can't really do anything about the exceptions during shutdown.

Restore to old scheme that posts shutdown to handler, and make
sure exceptions are handled.

Add a test for closing without start.

bug: 168660257
test: connectedCheck on Pixel4
Change-Id: I2f6912b5f8e01b6d2c78a363e85cbb907dce751d