Add ViewModelComponent and ViewModelScoped to androidx.hilt.lifecycle.
The ViewModelComponent serves as the source of all
@ViewModelInject-annotated ViewModels. It is a subcomponent
of the ActivityRetainedComponent that is seeded with the
SavedStateHandle for the ViewModel being created. It is a
single-use subcomponent that allows users to use SavedStateHandle
in its bindings which also allows extracting activity and fragments
parameters. Furthermore it removes the need for androidx.hilt
defining its own assisted factory since the subcomponent itself
will act as the factory for the ViewModel.

Test: androidx.hilt.lifecycle tests.
Relnote: Add APIs for scoping ViewModel dependencies provided by Hilt.
Change-Id: I5421576b2cada1eb431110f4845a9d2fafc7ce5e