Auto lookup of constant names in Compose
In the antipitation that most constants are defined in the same file
that have contains the runtime type of the constant instance:

Given an arbitrary instance investigate these classes:
- the instance class
- all implemented interfaces
- all super classes
- all enclosed classes

For:
- companion object values
- object instance values
- file level values

With this change FontWeight are now displayed as their weight value
i.e. W700 instead of Bold. This is because both constants are given
in FontWeight but Bold is sorted before W700 in the result from the
reflection API.

Bug: 169275057
Test: Unit test updated
Change-Id: Ibd86d53dd25a7cf489918b3cc64277324206889f