Fix applying nested snapshots
When a nested snapshot applies it is made visible in its parent
snapshot by removing its id from the parents set of invalid
snapshots thus making all state changes in the nested snapshot
valid and visible to the parent. This only works if the parent
snapshot has an id greater than the parent's id as all id's
greater than a snaphshot's id are considered invalid. This is
initially true for a nested snapshot but will not be true if it
is advanced such as when a nested snapshot is taken of it.

To correctly apply a nested snapshot the parent snapshot needs
to advance to an id greater than the snapshot being applied.

This triggered an issue reported in b/169629547 where the
focus state of a text field was created during subcomposition
but was not acessible after the subcomposition applied.

Test: SnapshotTests.aParentSnapshotCanAccessAStatObjectedCreateByANestedSnapshot
Fixes: b/169629547
Change-Id: I5173b0ee86d5dbebc43b9f7a464ea80e086927ac