Implement findPrimaryConstructor in KspTypeElement
This CL implements findPrimaryConstructor in KspTypeElement.

I've also found a bug in KotlinMetadata implementation where
we would crash if it does not have a primary contructor.

I've also added tests for Java sources which required adding
a new synthetic constructor element for Java classes which
are not abstract but has no explicit constructor.
https://github.com/google/ksp/issues/98

Also discovered another bug w/ KSP java interop where it could
return the same function declaration twice if it is coming from
java, added a workaround for it.

https://github.com/google/ksp/issues/99

With this change, we can enable more of XProcessingEnv tests to
include KSP. I've enabled all except for the primitive types test.

Bug: 160322705
Test: KspTypeElementTest#constructors, XProcessingEnvTest
Change-Id: I2def5651fb3d7d693e7256646168d12622eb6928