First version of the Ongoing Activities library.
Ongoing activities refer to user initiated & time bound activities that
could be running in the background of the watch (workouts, calls, media,
etc). This is a new API for handling ongoing activities that provides a
rich experience for the developer, using Notifications as a
communication mechanism but leaving the door open to change that in the
future.

Relnote: New API for Ongoing Activities, this is a no-op on "unsupported devices."
Test: Manual - automated tests on a follow up cl

Change-Id: I4f85e640fcd7ff0ff1b09f1dcbd68c8d6b1324e7