Make the AppSearchBackend interface async.
In R, our guidance was to make the framework API sync to make it easy to
reuse in androidx, which provides its own threading and ListenableFuture.

In S, our guidance is to make the framework API async. This means the
backend interface becomes more complex and resembles AppSearchManager itself.

Bug: 169377054
Test: AppSearchManagerTest
Relnote: N/A
Change-Id: I33ea506fe1a14f047c2ec488e9a4cade18158a2d