Introduce XSuspendMethodType
This CL removes getSuspendMethodReturnType from XMethodType and
instead moves it to a new XSuspendMethodType.

This change ensure that this method can only be called if the
caller ensured that this is a suspend method. It is safer and avoids
needing to throw an illegal state exception when called on non-suspend
methods.

Bug: 170258558
Test: existing tests
Change-Id: Ie59aa4b04a3789353b67efc98b9195b6ece042a8