Fix for the root node not being placed after it is reattached to the window
This resulted in compose views not being drawn at all if it was once detached breaking use cases similar to using ComposeView inside RecyclerView.
Also according to the documentation there are no guarantees about that and it is theoretically possible that our AndroidComposeView will be measured before receiving onAttachedToWindow callback. Our system is not ready for starting measuring LayoutNodes which are not attached, so I just guarded this with ifs. So if it happens we will remeasure again after being attached.

Fixes: 170446637
Test: manually, new tests in MeasureAndLayoutDelegateTest and AndroidLayoutDrawTest
Change-Id: Ie6cd429eaa5d0e73944daa7ec33ebe1a4e68e8fb
(cherry picked from commit 1dec8f9d587ffe9ac210406d3c3b8989454f3f05)