Messager implementation for KSP
I couldn't add a test for KSP because right now logging
errors do not fail compilation in KSP.
If by the time we need to run room compiler's failure tests,
if it is still not implemented in KSP, I'll update the
abstraction to track messages and throw error at the end
of compilation.

https://github.com/google/ksp/issues/122.

I've added a test to be enabled once that issue is fixed.

Bug: 160322705
Test: XProcessingEnvTest
Change-Id: Ied320bf329e026a71cc8c107f40e371e6bf8691e