Reload device quirks in tests that require device workarounds
Device quirks are loaded the first time a device workaround is encountered, and remain in memory until the process is killed. For tests, this means that the same device quirks are used for all of them. This causes an issue when tests modify device properties (using roboletric for instance). Instead of force-reloading the device quirks in every test that uses a device workaround, This CL makes DeviceQuirks internally reload the quirks every time a device workaround calls its method.

Bug: 170646523
Test: ImageCaptureOptionUnpackerTest.java
Change-Id: I50ba5dc375e7fb5e6f78082e2a5b5ca535bd0b87